% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_26.R
\name{bagging_classification}
\alias{bagging_classification}
\title{Bagging für Klassifikation}
\usage{
bagging_classification(X, y, B = 100, ...)
}
\arguments{
\item{Parameter:}{X: Matrix oder data.frame mit Features (n x d)
y: Vektor mit Klassenlabels (Länge n)
B: Anzahl Bootstrap-Samples (Standard: 100)
...: Weitere Parameter für fit_greedy_cart_classification (max_splits, min_leaf_size,...)}
}
\value{
Rückgabe:
  Objekt der Klasse "bagging_classification" mit:
    - trees: Liste der B trainierten Bäume
    - B: Anzahl der Bäume
    - levels: Klassenlabels
}
\description{
Bagging für Klassifikation
}
