% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_28.R
\name{bagging_classification_proba}
\alias{bagging_classification_proba}
\title{bagging classification probability}
\usage{
bagging_classification_proba(X, y, B = 100, ...)
}
\description{
bagging classification probability
}
