% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging_greedycart}
\alias{bagging_greedycart}
\title{Generating Bootstrapsamples}
\usage{
bagging_greedycart(data, n_bootstrapSamples, properties, n_nodes, mode, target)
}
\arguments{
\item{data}{is the dataset you are working on}

\item{n_bootstrapSamples}{number of bootstrap samples taken out of data nodes}

\item{properties}{number of properties}

\item{n_nodes}{number of nodes taken out of dataset starting from 1}

\item{mode}{select between regression and classification}

\item{target}{target attribute for tree generation}
}
\value{
generated trees
}
\description{
This function generates bootstrap samples based on input.
}
\examples{
bagging_greedycart
(
data = ames,
n_bootstrapSamples = 20,
properties = 20,
n_nodes = 100,
mode = "regression",
target = "Sale_Price"
)
}
