% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_25.R
\name{bagging_regression}
\alias{bagging_regression}
\title{Bagging für Regression}
\usage{
bagging_regression(X, y, B = 100, ...)
}
\arguments{
\item{Parameter:}{X: Matrix oder data.frame mit Features (n x d)
 y: Vektor mit Zielwerten (Länge n)
 B: Anzahl Bootstrap-Samples (Standard: 100)
...: Weitere Parameter für fit_greedy_cart_regression (max_splits, min_leaf_size,...)}
}
\value{
Rückgabe:
  Objekt der Klasse "bagging_regression" mit:
    - trees: Liste der B trainierten Bäume
    - B: Anzahl der Bäume
}
\description{
- Ziehe B Bootstrap-Samples (mit Zurücklegen)
- Baue auf jedem Sample einen Regressions-Baum
- Finale Vorhersage = Mittelwert aller Baum-Vorhersagen
}
