% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{choose_tp_lambda}
\alias{choose_tp_lambda}
\title{Choose Optimal Subtree for Given \eqn{\lambda}}
\usage{
choose_tp_lambda(lambda, pruning_sequence)
}
\arguments{
\item{lambda}{The complexity parameter for pruning.}

\item{pruning_sequence}{A list of subtrees generated during the pruning process.}
}
\value{
The subtree that minimizes the cost-complexity criterion.
}
\description{
Selects the subtree from the pruning sequence that minimizes the 
cost-complexity criterion for a given \eqn{\lambda}.
}
