% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{costComplexityPruning}
\alias{costComplexityPruning}
\title{this is an example documentation}
\usage{
costComplexityPruning(x)
}
\arguments{
\item{x}{whatever}
}
\description{
there are more tags...
compile with devtools::document()
}
