% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{CV}
\alias{CV}
\title{Cross-Validation Error Calculation}
\usage{
CV(lambda, sequences, Im, training_data_x, training_data_y)
}
\arguments{
\item{lambda}{The complexity parameter for pruning.}

\item{sequences}{A list of pruning sequences generated from different training data.}

\item{Im}{The partition indices for cross-validation.}

\item{training_data_x}{The training data x.}

\item{training_data_y}{The training data y.}
}
\value{
The cross-validation error for the given \eqn{\lambda}.
}
\description{
Calculates the cross-validation error for a given complexity parameter \eqn{\lambda} 
by pruning a decision trees with different training data and evaluating its performance on validation sets.
}
