% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{cv_optimal_lambda}
\alias{cv_optimal_lambda}
\title{Aus Bemerkung 6.21: M-Fold Cross-Validation
cross validation optimal lambda}
\usage{
cv_optimal_lambda(
  X,
  y,
  fit,
  seq_full,
  mode,
  M = 5,
  max_splits = .Machine$integer.max
)
}
\description{
Aus Bemerkung 6.21: M-Fold Cross-Validation
cross validation optimal lambda
}
