% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GierigesVerf_Regression.R
\name{fit_greedy_cart_regression}
\alias{fit_greedy_cart_regression}
\title{Fit greedy CART regression tree}
\usage{
fit_greedy_cart_regression(
  X,
  y,
  max_splits = 10^9,
  min_leaf_size = 1,
  min_improve = 1e-12,
  print_splits = TRUE,
  mtry = NULL
)
}
\description{
Fit greedy CART regression tree
}
