% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{get_optimal_tree_for_lambda}
\alias{get_optimal_tree_for_lambda}
\title{get optimal tree for lambda}
\usage{
get_optimal_tree_for_lambda(pruning_seq, lambda, y_train, mode)
}
\description{
get optimal tree for lambda
}
