% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{get_pruning_sequence}
\alias{get_pruning_sequence}
\title{Generate Pruning Sequence}
\usage{
get_pruning_sequence(t)
}
\arguments{
\item{t}{The original decision tree.}
}
\value{
A list of subtrees representing the pruning sequence (t(0), ..., t(p)).
}
\description{
Generates a sequence of subtrees using weakest link pruning from the original tree.
}
