% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{get_remaining_subtrees}
\alias{get_remaining_subtrees}
\title{Get Remaining Subtrees}
\usage{
get_remaining_subtrees(t0, possible_trees, remaining_trees)
}
\arguments{
\item{t0}{Indices of nodes belonging to subtree (relative to original tree)}

\item{possible_trees}{A list of all possible subtrees (their node indices in original tree).}
}
\value{
A vector of indices of trees in possible_trees that are subtrees of `t0`.
}
\description{
Filters the list of possible subtrees to include only those that are
subtrees of a specified tree.
}
