% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{get_sub_trees}
\alias{get_sub_trees}
\title{Generate Possible Subtrees}
\usage{
get_sub_trees(t)
}
\arguments{
\item{t}{The original decision tree.}
}
\value{
A list of nodes indices each forming a pruned subtree.
}
\description{
Generates all possible pruned subtrees of the original tree that share the same root.
}
