% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{Greedy algorithm}
\usage{
greedy(
  training_data_x,
  training_data_y,
  split_count = 10,
  data_in_leaves = 5,
  classification_tree = FALSE,
  random_subset = FALSE
)
}
\arguments{
\item{training_data_x}{Matrix of outputs (of given inputs)}

\item{split_count}{Max amount of split depth (0 for no limit)}

\item{data_in_leaves}{Amount data that should be in each leaf}

\item{classification_tree}{Should the tree be a classification tree?}

\item{random_subset}{Modification used for random_forest where only a random subset of dimensions is considered}
}
\description{
This function generates a decision tree with a greedy algorithm
}
