% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{L}
\alias{L}
\title{Loss Function}
\usage{
L(t, y, s)
}
\arguments{
\item{t}{The decision tree object.}

\item{y}{The true label.}

\item{s}{The predicted label.}
}
\value{
The computed loss value.
}
\description{
Computes the loss between the true label and the predicted label based on the type of tree (classification or regression).
}
\examples{
loss <- L(tree, y_true, y_pred)
}
