% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{make_partition}
\alias{make_partition}
\title{Create Cross-Validation Partitions}
\usage{
make_partition(x, m = 10)
}
\arguments{
\item{x}{The training data.}

\item{m}{The number of partitions to create. Default is 10.}
}
\value{
A vector indicating the partition assignments for each observation.
}
\description{
Creates a partition of the training data into `m` parts for cross-validation purposes.
}
