% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{new_tree}
\alias{new_tree}
\title{Generates new tree with two more children}
\usage{
new_tree(tree, index, k, max_k, min_data, A_parent_indices, m)
}
\arguments{
\item{tree}{Original tree}

\item{index}{Index of leaf}

\item{k}{Depth of Algorithm}

\item{max_k}{Max amount of split depth (0 for no limit)}

\item{min_data}{Amount data that should be in each leaf}

\item{A_parent_indices}{All the data classified in leaf node}

\item{m}{}
}
\description{
Generates new tree with two more children
}
