% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Klassifikation.R
\name{plot_classification_fit}
\alias{plot_classification_fit}
\title{Funktion zum Plotten eines einzelnen Klassifikationsbaums (2D)
plot classification fit}
\usage{
plot_classification_fit(tree, X, y, title, input_func = NULL, res = 200)
}
\description{
Funktion zum Plotten eines einzelnen Klassifikationsbaums (2D)
plot classification fit
}
