% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Regression.R
\name{plot_regression_fit}
\alias{plot_regression_fit}
\title{plot regression fit}
\usage{
plot_regression_fit(tree, X, y, title, original_function, show_original_func)
}
\description{
plot regression fit
}
