% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_26.R
\name{predict.bagging_classification}
\alias{predict.bagging_classification}
\title{Vorhersage-Funktion für Bagging-Klassifikation}
\usage{
\method{predict}{bagging_classification}(object, newdata, ...)
}
\description{
Implementiert Majority Vote aus Def 6.26:
f̂ₙᵇᵃᵍᵍ(x) = argmax_k #{b ∈ {1,...,B} : f̂ₙ*ᵇ(x) = k}
}
