% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_25.R
\name{predict.bagging_regression}
\alias{predict.bagging_regression}
\title{Vorhersage-Funktion für Bagging}
\usage{
\method{predict}{bagging_regression}(object, newdata, ...)
}
\arguments{
\item{Parameter:}{object: Bagging-Modell (von bagging_regression)
  newdata: Matrix/data.frame mit neuen Datenpunkten für Vorhersage
...: Weitere Parameter}
}
\value{
Rückgabe:
  Vektor mit Vorhersagen (Länge = Anzahl Zeilen in newdata)
}
\description{
Implementiert die Aggregation aus Def 6.25:
f̂ₙᵇᵃᵍᵍ(x) = (1/B) * Σᵇ₌₁ᴮ f̂ₙ*ᵇ(x)
}
