% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdm_Forest_Algo.R
\name{predict.random_forest_classification}
\alias{predict.random_forest_classification}
\title{Vorhersage-Funktion für Random Forests-Klassifikation}
\usage{
\method{predict}{random_forest_classification}(object, newdata, ...)
}
\description{
Vorhersage-Funktion für Random Forests-Klassifikation
}
