% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdm_Forest_Algo.R
\name{predict.random_forest_regression}
\alias{predict.random_forest_regression}
\title{Vorhersage-Funktion für Random Forests}
\usage{
\method{predict}{random_forest_regression}(object, newdata, ...)
}
\arguments{
\item{Parameter:}{object: Bagging-Modell (von bagging_regression)
newdata: Matrix/data.frame mit neuen Datenpunkten für Vorhersage
...: Weitere Parameter}
}
\value{
Rückgabe:
  Vektor mit Vorhersagen (Länge = Anzahl Zeilen in newdata)
}
\description{
Implementiert die Aggregation aus Def 6.25:
f̂ₙᵇᵃᵍᵍ(x) = (1/B) * Σᵇ₌₁ᴮ f̂ₙ*ᵇ(x)
}
