% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{predict_bagging}
\alias{predict_bagging}
\title{Predict with Bagging Models}
\usage{
predict_bagging(models, dataPoints, mode)
}
\arguments{
\item{models}{bagging models}

\item{dataPoints}{dataset with one or multiple rows}

\item{mode}{regression or classification}
}
\value{
vector of predictions
}
\description{
Predict with Bagging Models
}
