% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging_6_28.R
\name{predict_proba}
\alias{predict_proba}
\title{Definition 6.27: Klassenwahrscheinlichkeiten eines Baums}
\usage{
predict_proba(object, newdata, y_train)
}
\arguments{
\item{Parameter:}{object:  Baum-Objekt (greedy_cart_clas)
newdata: Matrix/data.frame mit neuen Datenpunkten
y_train: Trainings-Labels (factor), die zum Fitten benutzt wurden}
}
\value{
Rückgabe:
Matrix n_new x K mit Klassenwahrscheinlichkeiten
}
\description{
Traversiert den Baum wie predict, gibt aber pro Punkt
den Vektor p̂_k(A) = #{i : X_i ∈ A, Y_i = k} / #A zurück.
}
