% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForrest.R
\name{predict_rf}
\alias{predict_rf}
\title{Predict with a Random Forest Model}
\usage{
predict_rf(rf_model, newData)
}
\arguments{
\item{rf_model}{A list object returned by the \code{random_forest} function.}

\item{newData}{A data frame containing the features for which predictions
are to be made.}
}
\value{
A vector of predictions. Numerical values are returned for
  regression, and character/factor labels for classification.
}
\description{
Generates predictions for new data using a trained Random Forest model.
For regression, it averages the predictions of all trees. For classification,
it uses majority voting.
}
