% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_datastructure.R
\name{print.Tree}
\alias{print.Tree}
\title{Shows the type, length, depth and dimension of a tree.}
\usage{
\method{print}{Tree}(x)
}
\arguments{
\item{x}{tree}
}
\description{
Shows the type, length, depth and dimension of a tree.
}
