% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdm_Forest_Algo.R
\name{random_forest_classification}
\alias{random_forest_classification}
\title{Hauptfunktion: Random Forests für Klassifikation}
\usage{
random_forest_classification(X, y, B = 100, mtry = NULL, A_n = NULL, ...)
}
\arguments{
\item{Parameter:}{X: Matrix oder data.frame mit Features (n x d)
y: Vektor mit Klassenlabels (Länge n)
B: Anzahl Bootstrap-Samples (Standard: 100)
mtry: Anzahl der zufällig ausgewählten Features pro Split
A_n: Größe des Bootstrap-Samples (Standard: n)
...: Weitere Parameter für fit_greedy_cart_classification (max_splits, min_leaf_size,...)}
}
\value{
Rückgabe:
  Objekt der Klasse "random_forest_classification" mit:
    - trees: Liste der B trainierten Bäume
    - B: Anzahl der Bäume
    - levels: Klassenlabels

random forest for classification using greedy CART
}
\description{
Hauptfunktion: Random Forests für Klassifikation
}
