% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdm_Forest_Algo.R
\name{random_forest_regression}
\alias{random_forest_regression}
\title{Hauptfunktion: Random Forests für Regression}
\usage{
random_forest_regression(X, y, B = 100, mtry = NULL, A_n = NULL, ...)
}
\arguments{
\item{Parameter:}{X: Matrix oder data.frame mit Features (n x d)
y: Vektor mit Zielwerten (Länge n)
B: Anzahl Bootstrap-Samples (Standard: 100)
...: Weitere Parameter für fit_greedy_cart_regression (max_splits, min_leaf_size,...)}
}
\value{
Rückgabe:
  Objekt der Klasse "random_forest_regression" mit:
    - trees: Liste der B trainierten Bäume
    - B: Anzahl der Bäume

random forest for regression using greedy CART
}
\description{
- Ziehe B Bootstrap-Samples (mit Zurücklegen)
- Baue auf jedem Sample einen Regressions-Baum
- Finale Vorhersage = Mittelwert aller Baum-Vorhersagen
}
