% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costComplexityPruning.R
\name{rec_sub_trees}
\alias{rec_sub_trees}
\title{Recursive Subtree Generation}
\usage{
rec_sub_trees(t, t_sub, marked_nodes)
}
\arguments{
\item{t}{The original decision tree.}

\item{t_sub}{The vector of nodes belonging to current subtree in the generation process.}
}
\value{
A list of node indieces of subtrees generated from `t`.
}
\description{
Should not be used by itself!
}
\details{
A helper function that recursively generates all possible subtrees of a given tree.
}
