% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{show_results}
\alias{show_results}
\title{Visualise testresults}
\usage{
show_results(test_results)
}
\arguments{
\item{test_results}{results for some tests done prior}
}
\value{
prints and plots some testresult stats
}
\description{
Visualise testresults
}
\examples{
show_results(test_results)
}
