% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_datastructure.R
\name{[.Tree}
\alias{[.Tree}
\title{Allows accessing the n-th element of the tree (equivalent to array representation)
getting child and parents elements, determining whether}
\usage{
\method{[}{Tree}(x, value)
}
\arguments{
\item{x}{tree}

\item{value}{index of element}
}
\description{
Allows accessing the n-th element of the tree (equivalent to array representation)
getting child and parents elements, determining whether
}
