% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_datastructure.R
\name{[<-.Tree}
\alias{[<-.Tree}
\title{Allows writing to the the n-th element of the tree.
Vectorised mass assignment possible.
Passing NA as a value will result in the deletion of a node and all its children.}
\usage{
\method{[}{Tree}(tree, index) <- value
}
\arguments{
\item{tree}{tree}

\item{index}{index of element}

\item{value}{atomic vector of two real numbers (j, s)}
}
\description{
Allows writing to the the n-th element of the tree.
Vectorised mass assignment possible.
Passing NA as a value will result in the deletion of a node and all its children.
}
