% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{test_bagging}
\alias{test_bagging}
\title{Testing Bagging Results}
\usage{
test_bagging(models, dataPoints, mode, target)
}
\arguments{
\item{models}{bagging models}

\item{dataPoints}{test dataset}

\item{mode}{regression or classification}

\item{target}{target column}
}
\value{
data.frame with actual/predicted
}
\description{
Testing Bagging Results
}
\examples{
test_bagging(models = models, dataPoints = ames[1:10,], mode="regression",
target="Sale_Price")
}
