% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForrest.R
\name{test_rf}
\alias{test_rf}
\title{Test and Evaluate Random Forest Performance}
\usage{
test_rf(rf_model, test_data)
}
\arguments{
\item{rf_model}{The trained model from random_forest()}

\item{test_data}{Dataframe containing the test features and target}
}
\value{
A dataframe with actuals, predictions, and error metrics
}
\description{
Test and Evaluate Random Forest Performance
}
