% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_datastructure.R
\docType{class}
\name{Tree-class}
\alias{Tree-class}
\alias{Tree}
\title{Tree data structure. Allows accessing elements via index,
getting child and parents elements, determining whether 
an element is a leaf and finding the depth of an element.}
\arguments{
\item{length}{integer, is rounded to the next power of 2}

\item{dimension}{integer, dimension of the input data}
}
\description{
Tree data structure. Allows accessing elements via index,
getting child and parents elements, determining whether 
an element is a leaf and finding the depth of an element.
}

