% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging_greedycart}
\alias{bagging_greedycart}
\title{Bagging Greedy CART}
\usage{
bagging_greedycart(data, n_bootstrapSamples, properties, n_nodes, mode, target)
}
\arguments{
\item{data}{dataset}

\item{n_bootstrapSamples}{number of bootstrap samples}

\item{properties}{number of properties}

\item{n_nodes}{number of nodes}

\item{mode}{regression or classification}

\item{target}{target column}
}
\value{
list of trees
}
\description{
Bagging Greedy CART
}
\examples{
bagging_greedycart(data = ames, n_bootstrapSamples = 20, properties = 20,
n_nodes = 100, mode = "regression", target = "Sale_Price")
}
