% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForests.R
\name{random_forest}
\alias{random_forest}
\title{Random Forest}
\usage{
random_forest(training_data_x, training_data_y, classification_tree = FALSE)
}
\arguments{
\item{training_data_x}{Matrix of outputs (of given inputs)}

\item{classification_tree}{Should the tree be a classification tree?}
}
\value{
decision_function Decision Function used to predict output based on a given input
}
\description{
This function creates a decision function based on the random forest algorithm
}
