% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_datastructure.R
\name{[.Tree}
\alias{[.Tree}
\title{Allow accessing the n-th element of the tree (equivalent to array representation)
getting child and parents elements, determining whether}
\usage{
\method{[}{Tree}(tree, value)
}
\arguments{
\item{tree}{tree}

\item{value}{index of element}
}
\description{
https://www.geeksforgeeks.org/array-representation-of-binary-heap/
}
