% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{test_bagging}
\alias{test_bagging}
\title{Testing Bagging results}
\usage{
test_bagging(models, dataPoints, mode, target)
}
\arguments{
\item{models}{input trees from bagging_greedycart}

\item{dataPoints}{test datapoints}

\item{mode}{select between regression and classification}

\item{target}{target attribute}
}
\value{
generated trees
}
\description{
This function tests inputs on given bagging trees.
}
\examples{
test_bagging
(
models = models,
dataPoints = ames[120:130,],
mode = "regression",
target = "Sale_Price"
)
}
